requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-huggingface
faiss-cpu
pandas
sentence-transformers
tqdm
google-generativeai